requirements.txt
chess==1.10.0
stockfish==3.28.0
SpeechRecognition==3.10.0
pyttsx3==2.90
mysql-connector-python==8.2.0